Complete the leagues/list view page
The page consists of 3 justified tabs representing the leagues owned by
the user, those the user belongs to but does not own and those that the
user does not belong to and does not own.
Each tab displays the list of leagues in a list group where the list
group heading is the league name which is a link to the league/view
page. The list group item also presents a badge that shows the number of
teams in the league. The list group item also has a button for joining
or managing the league. And finally, the list group item presents the
username of the league owner underneath the league name.