Completed the leagues/view  view page
This page lists the teams in the league in a list group view. It also
shows a warning if the user is not a member of the league and if the
user is a member but has not registered a team yet in which case the
page presents a button for registering a new team in the league.

If the user is the league owner, the page presents him with buttons to
edit the league, delete the league and the invite code for the league as
well as a button to copy the invite code to the clipboard.