Complete the teams/list view page
Presents each team owned by the user as a list group item. The list
group is also presented as a thumbnail gallery with each list group item
displaying the team logo as an image and the caption as a combination of
the team name and the name of the league the team belongs to.